conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.7.1=pyhd8ed1ab_0
aom=3.5.0=h7ea286d_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py311he2be06e_3
asttokens=2.2.1=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pyha770c72_0
bleach=6.0.0=pyhd8ed1ab_0
blosc=1.21.4=hc338f07_0
bokeh=3.2.0=pyhd8ed1ab_0
brotli=1.0.9=h1a8c8d9_9
brotli-bin=1.0.9=h1a8c8d9_9
brotli-python=1.0.9=py311ha397e9f_9
brunsli=0.1=h9f76cd9_0
bzip2=1.0.8=h3422bc3_4
c-ares=1.19.1=hb547adb_0
c-blosc2=2.10.0=h068da5f_0
ca-certificates=2023.12.12=hca03da5_0
certifi=2023.11.17=py311hca03da5_0
cffi=1.15.1=py311hae827db_3
cfitsio=4.2.0=h2f961c4_0
charls=2.4.2=h13dd4ca_0
charset-normalizer=3.1.0=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
cloudpickle=2.2.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
colorcet=3.0.1=pyhd8ed1ab_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.1.0=py311he4fd1f5_0
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.12.0=py311he2be06e_1
dask-core=2023.6.1=pyhd8ed1ab_0
datashader=0.15.1=pyhd8ed1ab_0
datashape=0.5.4=py_1
dav1d=1.2.1=hb547adb_0
debugpy=1.6.7=py311ha397e9f_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.2=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
flit-core=3.9.0=pyhd8ed1ab_0
fonttools=4.40.0=py311heffc1b2_0
freetype=2.12.1=hd633e50_1
fsspec=2023.6.0=pyh1a96a4e_0
giflib=5.2.1=h1a8c8d9_3
holoviews=1.16.2=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
imagecodecs=2023.1.23=py311ha72d3b9_1
imageio=2.31.1=pyh24c5eb1_0
importlib-metadata=6.7.0=pyha770c72_0
importlib_metadata=6.7.0=hd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
ipykernel=6.23.1=pyh736e0ef_0
ipython=8.14.0=pyhd1c38e8_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.7=pyhd8ed1ab_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.0=pyhd8ed1ab_1
jsonschema=4.18.0=pyhd8ed1ab_0
jsonschema-specifications=2023.6.1=pyhd8ed1ab_0
jupyter=1.0.0=py311h267d04e_8
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h267d04e_0
jupyter_events=0.6.3=pyhd8ed1ab_0
jupyter_server=2.7.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.8=pyhd8ed1ab_0
jxrlib=1.1=h27ca646_2
kiwisolver=1.4.4=py311hd6ee22a_1
krb5=1.20.1=h69eda48_0
lazy_loader=0.2=pyhd8ed1ab_0
lcms2=2.15=hd835a16_1
lerc=4.0.0=h9a09cb3_0
libaec=1.0.6=hb7217d7_1
libavif=0.11.1=h9f83d30_2
libblas=3.9.0=17_osxarm64_openblas
libbrotlicommon=1.0.9=h1a8c8d9_9
libbrotlidec=1.0.9=h1a8c8d9_9
libbrotlienc=1.0.9=h1a8c8d9_9
libcblas=3.9.0=17_osxarm64_openblas
libcurl=8.1.2=h912dcd9_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=12_2_0_hd922786_31
libgfortran5=12.2.0=h0eea778_31
libjpeg-turbo=2.1.5.1=h1a8c8d9_0
liblapack=3.9.0=17_osxarm64_openblas
libllvm14=14.0.6=hd1a9a77_3
libnghttp2=1.52.0=hae82a92_0
libopenblas=0.3.23=openmp_hc731615_0
libpng=1.6.39=h76d750c_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.42.0=hb31c410_0
libssh2=1.11.0=h7a5bd25_0
libtiff=4.5.0=hdc14d85_5
libwebp-base=1.3.1=hb547adb_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
libzopfli=1.0.3=h9f76cd9_0
linkify-it-py=2.0.0=pyhd8ed1ab_0
llvm-openmp=16.0.6=h1c12783_0
llvmlite=0.40.1=py311hea943cd_0
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.4=hb7217d7_0
markdown=3.4.3=pyhd8ed1ab_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.3=py311heffc1b2_0
matplotlib=3.7.1=py311ha1ab1f8_0
matplotlib-base=3.7.1=py311h99a5f44_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdit-py-plugins=0.4.0=pyhd8ed1ab_0
mdurl=0.1.0=pyhd8ed1ab_0
mistune=3.0.0=pyhd8ed1ab_0
multipledispatch=0.6.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=1.0.0=pyhb4ecaf3_1
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=7.6.0=pyhd8ed1ab_0
nbconvert-core=7.6.0=pyhd8ed1ab_0
nbconvert-pandoc=7.6.0=pyhd8ed1ab_0
nbformat=5.9.0=pyhd8ed1ab_0
ncurses=6.4=h7ea286d_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pyhd8ed1ab_0
notebook=6.5.4=pyha770c72_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numba=0.57.1=py311hbf3c4e2_0
numpy=1.24.4=py311hb8f3215_0
openjpeg=2.5.0=hbc2ba62_2
openssl=3.1.1=h53f4e23_1
overrides=7.3.1=pyhd8ed1ab_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py311h9e438b8_0
pandoc=3.1.3=hce30654_0
pandocfilters=1.5.0=pyhd8ed1ab_0
panel=1.2.0=pyhd8ed1ab_0
param=1.13.0=pyh1a96a4e_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.4.0=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=py311h095fde6_1
pip=23.1.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=3.8.0=pyhd8ed1ab_0
plotly=5.18.0=pypi_0
pooch=1.7.0=pyha770c72_3
prometheus_client=0.17.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=py311he2be06e_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyct=0.4.6=py_0
pyct-core=0.4.6=py_0
pygments=2.15.1=pyhd8ed1ab_0
pynndescent=0.5.10=pyh1a96a4e_0
pyobjc-core=9.2=py311hb702dc4_0
pyobjc-framework-cocoa=9.2=py311hb702dc4_0
pyparsing=3.1.0=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.4=h47c9636_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.21.0=py311hca03da5_0
python-fastjsonschema=2.17.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=3_cp311
pytz=2023.3=pyhd8ed1ab_0
pyviz_comms=2.3.2=pyhd8ed1ab_0
pywavelets=1.4.1=py311h4add359_0
pyyaml=6.0=py311he2be06e_5
pyzmq=25.1.0=py311hb1af645_0
readline=8.2=h92ec313_1
referencing=0.29.1=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.8.6=py311h0563b04_0
scikit-image=0.20.0=py311h4eec4a9_1
scikit-learn=1.3.0=py311hf0b18b8_0
scipy=1.11.1=py311h93d07a4_0
send2trash=1.8.2=pyhd1c38e8_0
setuptools=68.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
tbb=2021.9.0=hffc8910_0
tenacity=8.2.3=pypi_0
terminado=0.17.1=pyhd1c38e8_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2023.7.4=pyhd8ed1ab_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.2=py311heffc1b2_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023c=h71feb2d_0
uc-micro-py=1.0.1=pyhd8ed1ab_0
umap-learn=0.5.3=py311h267d04e_1
urllib3=2.0.3=pyhd8ed1ab_1
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.6.1=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
widgetsnbextension=4.0.8=pyhd8ed1ab_0
xarray=2023.6.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xyzservices=2023.5.0=pyhd8ed1ab_1
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zfp=1.0.0=hb6e4faa_3
zipp=3.15.0=pyhd8ed1ab_0
zlib-ng=2.0.7=h1a8c8d9_0
zstd=1.5.2=hf913c23_6
